Restructured for Min-Width
I think I possibly (emphasis on the possibly) understand how to structure with min-width now.  If you can comment on this being the proper way as opposed to the last commit, it would be appreciated.  Mixins --- reverted the min width variables to 400, 768, 1100.  Main-content --- Restructured the media queries/default structures of each element to what makes more sense to me (and is hopefully the proper way).